Remove diagram PDF concatenation; bump version to 0.9.5.3
Installing poppler-utils via apt (via execDepends) no longer works as of
mid-August, since poppler deprecated all of their prior releases after
finding some security bugs (as I understand it). Compiling the pdfunite
binary separately is a pain, and it's not that important for the diagram
PDFs to be concatenated into a single file anyway -- the PNG scatter
plots are no longer bundled either.